feat: Add placeholder support to EverCache
This commit adds support for a placeholder function in the EverCache class. The placeholder function can be used to provide a temporary value while the actual value is being fetched. If a placeholder function is provided, it will be called when the value is accessed before it has been fetched.

Co-authored-by: [Author Name]